Refactor App and Form components; remove FormularioPage. Updated routing structure and integrated new form functionalities with improved UI components.
